requirements.txt
# Core dependencies
feedparser==6.0.11
httpx==0.27.0
aiosqlite==0.20.0
pyyaml==6.0.1
python-dateutil==2.8.2

# ML/AI
sentence-transformers==2.5.1
scikit-learn==1.4.0
numpy==1.26.4

# Clustering
hdbscan==0.8.33

# Logging
structlog==24.1.0

# Utilities
python-dotenv==1.0.1
